requirements.txt
django==5.0.2
openai>=1.3.0
git+https://github.com/openai/swarm.git
sentence-transformers==2.2.2
python-dotenv==1.0.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
